Remove www bucket on S3
If we're able to redirect all S3 traffic from the www bucket
to the non-www bucket, this pull request can be accepted.

Please do not accept this pull request until the steps listed
in the "Amazon S3 Setup to Support WWW and Non-WWW" section
of the README have been completed.